requirements.txt
fastapi
uvicorn
pydantic>=2

langchain>=0.2
langchain-community>=0.2
# langchain-unstructured removed - using langchain_community UnstructuredFileLoader instead

pypdf
pymupdf
pdfplumber
unstructured
python-magic-bin; sys_platform == 'win32'
python-magic; sys_platform != 'win32'

sentence-transformers

langchain-openai
openai
pydantic-settings
python-dotenv
tiktoken
chromadb
